// src/app/enums/map-constants.enum.ts
export enum MapConstantsEnum {
    // Initial map settings
    LATITUDE_INIT = 0,
    LONGITUDE_INIT = 0,
    ZOOM_INIT = 3,
    ZOOM_MAX = 18,
    ZOOM_MIN = 3,
    ICON_SIZE_WIDTH = 25,
    ICON_SIZE_HEIGHT = 41,
    ICON_ANCHOR_WIDTH = 12,
    ICON_ANCHOR_HEIGHT = 41,
    POPUP_ANCHOR_X = 1,
    POPUP_ANCHOR_Y = -34,
    TOOLTIP_ANCHOR_X = 16,
    TOOLTIP_ANCHOR_Y = -28,
    SHADOW_SIZE_WIDTH = 41,
    SHADOW_SIZE_HEIGHT = 41,
}